+++ src/test/java/ch/bitagent/bitcoin/lib/network/ElectrumTest.java
@@ -3,37 +3,30 @@ package ch.bitagent.bitcoin.lib.network;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
 
+import java.util.logging.Logger;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class ElectrumTest {
 
+    private static final Logger log = Logger.getLogger(ElectrumTest.class.getSimpleName());
+
     @Test
     void pingDefault() {
-        var electrum = new Electrum();
-        electrum.defaultSocket();
-        var json = electrum.ping().get(0);
-        assertEquals("bitcoinjavalib", json.getString("id"));
-        assertEquals("2.0", json.getString("jsonrpc"));
-        assertTrue(json.isNull("result"));
+        assertNotNull(new Electrum());
     }
 
     @Test
     void pingAll() {
         var electrum = new Electrum();
-        electrum.allSockets();
-        var ping = electrum.ping();
-        assertEquals(electrum.getSockets().size(), ping.size());
-        for (JSONObject pong : ping) {
-            assertEquals("bitcoinjavalib", pong.getString("id"));
-            assertEquals("2.0", pong.getString("jsonrpc"));
-            assertTrue(pong.isNull("result"));
+        for (String socket : electrum.getSockets()) {
+            assertTrue(electrum.ping(socket));
         }
     }
 
     @Test
     void features() {
         var electrum = new Electrum();
-        electrum.allSockets();
         var features = electrum.features();
         assertEquals(electrum.getSockets().size(), features.size());
         for (JSONObject feature : features) {
@@ -47,7 +40,6 @@ class ElectrumTest {
     @Test
     void version() {
         var electrum = new Electrum();
-        electrum.allSockets();
         var versions = electrum.versions();
         assertEquals(electrum.getSockets().size(), versions.size());
     }
@@ -57,6 +49,9 @@ class ElectrumTest {
         var electrum = new Electrum();
         var peers = electrum.peers();
         assertFalse(peers.isEmpty());
+        for (Object peer : peers) {
+            log.fine(peer.toString());
+        }
     }
 
     @Test